// front-office-app/src/app/api/api-reference.ts
export const API_BASE_URL = new InjectionToken<string>('API_BASE_URL');
// ...
export interface IQuestionsClient {
    generateExam(command: GenerateExamCommand): Observable<GenerateExamResponse>;
}

@Injectable({
    providedIn: 'root'
})
export class QuestionsClient implements IQuestionsClient {
    private http: HttpClient;
    private baseUrl: string;
    protected jsonParseReviver: ((key: string, value: any) => any) | undefined = undefined;

    constructor(@Inject(HttpClient) http: HttpClient, @Optional() @Inject(API_BASE_URL) baseUrl?: string) {
        this.http = http;
        this.baseUrl = baseUrl ?? "https://localhost:44382";
    }

    generateExam(command: GenerateExamCommand): Observable<GenerateExamResponse> {
        let url_ = this.baseUrl + "/api/questions/random";
        url_ = url_.replace(/[?&]$/, "");

        const content_ = JSON.stringify(command);

        let options_ : any = {
            body: content_,
            observe: "response",
            responseType: "blob",
            headers: new HttpHeaders({
                "Content-Type": "application/json",
                "Accept": "application/json"
            })
        };

        return this.http.request("post", url_, options_).pipe(_observableMergeMap((response_ : any) => {
            return this.processGenerateExam(response_);
        })).pipe(_observableCatch((response_: any) => {
            if (response_ instanceof HttpResponseBase) {
                try {
                    return this.processGenerateExam(response_ as any);
                } catch (e) {
                    return _observableThrow(e) as any as Observable<GenerateExamResponse>;
                }
            } else
                return _observableThrow(response_) as any as Observable<GenerateExamResponse>;
        }));
    }

    protected processGenerateExam(response: HttpResponseBase): Observable<GenerateExamResponse> {
        const status = response.status;
        const responseBlob =
            response instanceof HttpResponse ? response.body :
            (response as any).error instanceof Blob ? (response as any).error : undefined;

        let _headers: any = {}; if (response.headers) { for (let key of response.headers.keys()) { _headers[key] = response.headers.get(key); }}
        if (status === 200) {
            return blobToText(responseBlob).pipe(_observableMergeMap((_responseText: string) => {
            let result200: any = null;
            let resultData200 = _responseText === "" ? null : JSON.parse(_responseText, this.jsonParseReviver);
            result200 = GenerateExamResponse.fromJS(resultData200);
            return _observableOf(result200);
            }));
        } else if (status !== 200 && status !== 204) {
            return blobToText(responseBlob).pipe(_observableMergeMap((_responseText: string) => {
            return throwException("An unexpected server error occurred.", status, _responseText, _headers);
            }));
        }
        return _observableOf(null as any);
    }
}
// ...
export class GenerateExamResponse implements IGenerateExamResponse {
    examId?: number;
    questions?: GenerateExamQuestionResponse[];

    constructor(data?: IGenerateExamResponse) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.examId = _data["examId"];
            if (Array.isArray(_data["questions"])) {
                this.questions = [] as any;
                for (let item of _data["questions"])
                    this.questions!.push(GenerateExamQuestionResponse.fromJS(item));
            }
        }
    }

    static fromJS(data: any): GenerateExamResponse {
        data = typeof data === 'object' ? data : {};
        let result = new GenerateExamResponse();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["examId"] = this.examId;
        if (Array.isArray(this.questions)) {
            data["questions"] = [];
            for (let item of this.questions)
                data["questions"].push(item.toJSON());
        }
        return data;
    }
}

export interface IGenerateExamResponse {
    examId?: number;
    questions?: GenerateExamQuestionResponse[];
}

export class GenerateExamQuestionResponse implements IGenerateExamQuestionResponse {
    id?: number;
    text?: string;
    category?: Category;
    type?: TypeButton;
    answers?: GenerateExamAnswerResponse[];

    constructor(data?: IGenerateExamQuestionResponse) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"];
            this.text = _data["text"];
            this.category = _data["category"];
            this.type = _data["type"];
            if (Array.isArray(_data["answers"])) {
                this.answers = [] as any;
                for (let item of _data["answers"])
                    this.answers!.push(GenerateExamAnswerResponse.fromJS(item));
            }
        }
    }

    static fromJS(data: any): GenerateExamQuestionResponse {
        data = typeof data === 'object' ? data : {};
        let result = new GenerateExamQuestionResponse();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id;
        data["text"] = this.text;
        data["category"] = this.category;
        data["type"] = this.type;
        if (Array.isArray(this.answers)) {
            data["answers"] = [];
            for (let item of this.answers)
                data["answers"].push(item.toJSON());
        }
        return data;
    }
}
// ...
export class GenerateExamCommand implements IGenerateExamCommand {
    candidateId?: number;

    constructor(data?: IGenerateExamCommand) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (<any>this)[property] = (<any>data)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.candidateId = _data["candidateId"];
        }
    }

    static fromJS(data: any): GenerateExamCommand {
        data = typeof data === 'object' ? data : {};
        let result = new GenerateExamCommand();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["candidateId"] = this.candidateId;
        return data;
    }
}

export interface IGenerateExamCommand {
    candidateId?: number;
}